Remove 1.x message-template code from 2.x
There are a few reasons.  The one which spurs action now -- I'd like to
simplify/align the permission models but don't want to rework the old
message-template code.

More broadly, it'll be easier to test/validate the main 2.x functionality if
this stuff isn't in-scope, and IMHO the ideal UI for managing
message-templates will require some reworking.