crmMosaicoListCtrl - Hide delete if not supported
The intent is to align with core's permissions. In core, the permission
to create/edit and delete are different, so we have to allow for
the possibility that user has one but not the other.

This commit is agnostic to what the permissions are -- e.g.  the UI takes
it's cue from the Mosaico API, and the Mosaico API takes its cue from the
MessageTemplate API.